// src/app/features/service-provider/service-providers/models/ServiceProviders.ts
import { ServiceProviderTypes } from '../../service-provider-types/models/ServiceProviderTypes';

export interface ServiceProviders {
  ServiceProviderId?: number; // optional because it's generated by the backend
  Name: string;
  VATNumber: string;
  CompanyRegNo: string;
  Branch: string;
  OfficeAddress: string;
  StorageAddress: string;
  TownCity: string;
  Province: string;
  ServiceProviderTypeId: number;
  DesignationNumber: string;
  Manager: string;
  AuthorisedBy: number;
  RatePerKm: number;
  RateAuthorisedOn: string; // ISO format expected
  IsActive: boolean;
  ActiveOn: string;
  ActiveBy: number;
  IsVerified: boolean;
  VerifiedOn: string;
  VerifiedBy: number;
  IsAccredited: boolean;
  AccreditedOn: string;
  AccreditedBy: number;
  IsDeleted?: boolean; // nullable in backend, so mark as optional
  ServiceProviderTypes?: ServiceProviderTypes | null; // can be null
  ContactDetails?: ContactDetail[]; // optional and nullable in backend
}

export interface ContactDetail {
  Id?: number; // optional, as it could be autogenerated or not sent on creation
  Type: string;
  Code: string;
  Value: string;
  Name: string;
  Comments: string;
}